Move modules into modules directory, fix imports too
