+++ main_window/modules/data_handlers.py
@@ -13,8 +13,8 @@ be updated, the appropriate signal is emitted from this class.
 import numpy as np
 from PySide6.QtCore import Signal, QObject, Slot
 
-from .utils import packet_spec
-from .plot_info import PlotInfo, PlotDataDisplayMode
+from ..utils import packet_spec
+from ..plot_info import PlotInfo, PlotDataDisplayMode
 
 class DataHandler(QObject):
