+++ main_window/utils/__init__.py
@@ -1,0 +1,3 @@
+from .csv_writer import CSVWriter
+from .data_conversions import *
+from .packet_spec import *
